feat: add pesawat, penerbangan, cari, filter, bonus feature

+++ sipes/src/main/java/apap/tugas/sipes/repository/PesawatDb.java
@@ -1,13 +1,14 @@
 package apap.tugas.sipes.repository;
 
 import apap.tugas.sipes.model.PesawatModel;
+import apap.tugas.sipes.model.TipeModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PesawatDb extends JpaRepository<PesawatModel, BigInteger>{
-    Optional<PesawatModel> findById(BigInteger id);
+public interface PesawatDb extends JpaRepository<PesawatModel, Long>{
+    Optional<PesawatModel> findById(Long id);
 }
